msm: smem: clarify SBL misconfiguration error messages
SMEM must be initialized by the secondary bootloader (SBL) before Linux
is booted and the SMEM driver is initialized.  If it is not initialized,
error messages are printed and SMEM is disabled.  Some of these error
messages can be misinterpreted to be a Linux configuration issue instead
of an SBL issue which delays resolving the misconfiguration.

Add additional code comments and change the error messages to make it
clear that the SBL has not initialized SMEM.

Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>