msm: smd: convert to secured smem feature
The shared memory (smem) driver provides a feature where point-to-point
smem items shared between Linux apps and one other remote processor are
protected against a number of attacks by security hardware, provided system
support exists.  Clients of the smem driver must elect to use this feature
if applicable.

SMD is a point-to-point communication protocol to which the security
feature applies.  Update SMD to make use of the security feature.

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>