msm: smsm: add snapshot queue scheduling delay
When debugging delayed SMSM callbacks, it is difficult to determine if
the delay is due to a scheduling issue for the callback worker or a
delayed interrupt issue.  In addition, sometimes the interrupt timestamp
needs to be determined from a crash dump.

Add timestamp when the SMSM interrupt snapshot is queued and calculate
the time delta when the snapshot is dequeued.

Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>