msm: vidc: Advertise extradata size in queue_setup()
If extradata is enabled, the correct size must be advertised as part of
queue_setup().  Prior to this commit, the output buffer size was taken
to be the extradata size.

In order to make sure that we aren't advertising stale buffer sizes,
query the core for both CAPTURE and OUTPUT planes.

Signed-off-by: Pushkaraj Patil <ppatil@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>